+++ app/Views/index.php
@@ -1,0 +1,9 @@
+  <!-- Admin Login -->
+  <a href="<?= url('/admin/login') ?>">
+    admin login
+  </a>
+
+  <!-- Karyawan Login-->
+  <a href="<?= url('/karyawan/login') ?>">
+    karyawan login
+  </a>